harness: collapse 9 files into 5 — policy / substrate / providers / recurrent / __init__
Structure is now isomorphic to the concepts it carries:

- policy.py absorbs constants.py + events.py + router.py. The Router
  class is kept as a thin @dataclass wrapper whose classify() delegates
  to Policy.classify() — same object classifies the turn it configures.
- substrate.py is prompt.py renamed. The file was never about a prompt.
  It is the layered substrate the model stands on: identity, context,
  retrieval, invocation.
- providers.py absorbs tools.py, including INTROSPECT_TOOL_SPEC
  (previously an uncommitted Spark diff). ToolSpec + BASH/DELEGATE/
  INTROSPECT specs sit above absorb_gate, validate_command, BashTool,
  and the provider protocol with Anthropic/OpenAI/Registry.
- recurrent.py: one-line import update (.prompt -> .substrate).
- __init__.py: full re-export surface preserved so downstream callers
  and tests that import by symbol name keep working across the old
  file names.

External callers updated: vybn_spark_agent.py and four test files
switch from harness.tools/harness.prompt/harness.router to
harness.providers/harness.substrate/harness.policy. test_live_repl_fixes
drops the INTROSPECT monkey-patch shim — providers.py owns the spec.

Pre-existing test failures (default_role, absent vybn_chat_api.py)
are unchanged by this commit: same 6 failures / 15 errors with or
without the collapse. No new breakage.

decide() is shipped upstream in vybn-phase PR #4. Wiring it into
Policy.classify() is a follow-up — this commit is the structural
collapse that makes that hookup a two-line change.
